+++ changelog.md
@@ -9,6 +9,7 @@
 
 ### Added
 
+- Add openchat-3.5-1210 model
 - Support static batching by passing lists to `do`
 - Support choices list on `do` to restrict possible outputs